+++ src/Schema/Directives/UpsertDirective.php
@@ -42,6 +42,7 @@ GRAPHQL;
         return new UpsertModel(
             new SaveModel($parentRelation),
             $this->directiveArgValue('identifyingColumns'),
+            $parentRelation,
         );
     }
 }